+++ src/__fixtures__/forms/dynamicForm.ts
@@ -3,7 +3,6 @@ import { ComponentWrapper } from "../../component-wrapper/ComponentWrapper";
 import { DynamicComponent } from "../../component-wrapper/DynamicComponent";
 import { StaticComponent } from "../../component-wrapper/StaticComponent";
 import { NamedSchema } from "../../database/types";
-import { StepDefinition } from "../../step/StepDefinition";
 
 import { SimpleSubmit } from "../components/SimpleSubmit";
 import { TestDynamicComponent } from "../components/TestDynamicComponent";
@@ -19,14 +18,14 @@ export type DynamicFormSchema = NamedSchema<
   }
 >;
 
-export const dynamicForm: { steps: StepDefinition[] } = {
+export const dynamicForm = {
   steps: [
     {
       slug: "test-step-1",
       nextSlug: "test-step-2",
       Submit: SimpleSubmit,
       componentWrappers: [
-        ComponentWrapper.wrapStatic(
+        ComponentWrapper.wrapStatic<{ src?: string }, DynamicFormSchema>(
           new StaticComponent({
             key: "test-img",
             Component: "img",
@@ -48,7 +47,8 @@ export const dynamicForm: { steps: StepDefinition[] } = {
             >({
               storeName: "testStore",
               key: 0
-            })
+            }),
+            defaultValue: "test default"
           })
         )
       ]
@@ -58,7 +58,7 @@ export const dynamicForm: { steps: StepDefinition[] } = {
       nextSlug: "test-step-1",
       Submit: SimpleSubmit,
       componentWrappers: [
-        ComponentWrapper.wrapStatic(
+        ComponentWrapper.wrapStatic<{ src?: string }, DynamicFormSchema>(
           new StaticComponent({
             key: "test-img-2",
             Component: "img",
@@ -80,7 +80,8 @@ export const dynamicForm: { steps: StepDefinition[] } = {
             >({
               storeName: "testStore",
               key: 1
-            })
+            }),
+            defaultValue: "test default"
           })
         )
       ]